requirements.txt
attrs==21.4.0
iniconfig==1.1.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tomli==2.0.1